backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
boto3==1.42.9
botocore==1.42.9
Django==5.2.8
django-cors-headers==4.9.0
django-ninja==1.5.1
djangorestframework==3.16.1
jmespath==1.0.1
numpy==2.3.5
pgvector==0.4.2
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
